+++ Projects/Coffice/Sources/App/AppDelegate.swift
@@ -9,21 +9,16 @@
 import Firebase
 import KakaoSDKAuth
 import KakaoSDKCommon
+import KakaoSDKUser
 import UIKit
 
 class AppDelegate: NSObject, UIApplicationDelegate {
+
   func application(
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
   ) -> Bool {
-    setUpKakaoLogin()
     FirebaseApp.configure()
     return true
-  }
-
-  private func setUpKakaoLogin() {
-      guard let kakaoNativeAppKey = CofficeResources.bundle
-      .object(forInfoDictionaryKey: "KAKAO_NATIVE_APP_KEY") as? String else { return }
-    KakaoSDK.initSDK(appKey: kakaoNativeAppKey)
   }
 }
